requirements.txt
aiogram==2.25.1
python-dotenv==0.19.0
sqlalchemy==1.4.36
psycopg2-binary==2.9.3
python-binance==1.0.19
